Potential fix for code scanning alert no. 34: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using the larger type size_t to avoid overflow. This can be achieved by casting one of the operands to size_t before performing the multiplication. This way, the multiplication will be done in the larger type, preventing overflow.

 - Cast one of the operands (dec->width_ or dec->height_) to size_t before the multiplication.
 - This change should be made on line 96 where the multiplication occurs.

+++ libwebp-1.0.3/src/dec/alpha_dec.c
@@ -93,7 +93,7 @@ static int ALPHInit(ALPHDecoder* const dec, const uint8_t* data,
   // No need to copy the scaling parameters.
 
   if (dec->method_ == ALPHA_NO_COMPRESSION) {
-    const size_t alpha_decoded_size = dec->width_ * dec->height_;
+    const size_t alpha_decoded_size = (size_t)dec->width_ * dec->height_;
     ok = (alpha_data_size >= alpha_decoded_size);
   } else {
     assert(dec->method_ == ALPHA_LOSSLESS_COMPRESSION);
